% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{curl_echo}
\alias{curl_echo}
\title{Echo Service}
\usage{
curl_echo(handle, port = 9359)
}
\arguments{
\item{handle}{a curl handle object}

\item{port}{the port number on which to run httpuv server}
}
\description{
This function is only for testing purposes. It starts a local httpuv server to
echo the request body and content type in the response.
}
\examples{
h <- handle_setform(new_handle(), foo = "blabla", bar = charToRaw("test"),
myfile = form_file(system.file("DESCRIPTION"), "text/description"))
req <- curl_echo(h)

# Show the multipart body
cat(rawToChar(req$content))
}
