% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{curl_getdate}
\alias{curl_getdate}
\title{Parse date/times.}
\usage{
curl_getdate(datestring)
}
\arguments{
\item{datestring}{a string consisting of a timestamp}
}
\description{
Parse date/times.
}
\examples{
# Parse dates in many formats
curl_getdate("Sunday, 06-Nov-94 08:49:37 GMT")
curl_getdate("06 Nov 1994 08:49:37")
curl_getdate("20040911 +0200")
}

