% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/curl_perform.R
\name{curl_perform}
\alias{curl_perform}
\title{Perform a request}
\usage{
curl_perform(url)
}
\arguments{
\item{url}{A character string naming the URL of a resource to be downloaded.}
}
\description{
Perform a request
}
\examples{
# Redirect + cookies
res <- curl_perform("http://httpbin.org/cookies/set?foo=123&bar=ftw")
}

