% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{form_file}
\alias{form_file}
\title{Upload files with form}
\usage{
form_file(path, type = NULL)
}
\arguments{
\item{path}{a string with a path to an existing file on disk}

\item{type}{MIME content-type of the file.}
}
\description{
Upload files with form
}

