% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_cookies}
\alias{handle_cookies}
\title{Extract cookies from a handle}
\usage{
handle_cookies(handle)
}
\arguments{
\item{handle}{a curl handle object}
}
\description{
Extract cookies from a handle
}
\examples{
h <- new_handle()
handle_cookies(h)

# Server sets cookies
req <- curl_perform("http://httpbin.org/cookies/set?foo=123&bar=ftw", handle = h)
handle_cookies(h)

# Server deletes cookies
req <- curl_perform("http://httpbin.org/cookies/delete?foo", handle = h)
handle_cookies(h)
}
\seealso{
Other handle.functions: \code{\link{handle_reset}},
  \code{\link{handle_setform}},
  \code{\link{handle_setheaders}},
  \code{\link{handle_setopt}}; \code{\link{new_handle}}
}

