% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_setopt}
\alias{handle_reset}
\alias{handle_setform}
\alias{handle_setheaders}
\alias{handle_setopt}
\title{Set handle options.}
\usage{
handle_setopt(handle, ...)

handle_setheaders(handle, ...)

handle_setform(handle, ...)

handle_reset(handle)
}
\arguments{
\item{handle}{Handle to modify}

\item{...}{additional named options / headers to be set in the handle.
To send a file, see \code{\link{form_file}}. To list all allowed options,
see \code{\link{curl_options}}}
}
\value{
All functions modify the handle in place but also return the handle
  so you can create a pipeline of operations.
}
\description{
Currently \code{handle_setopt} will append options where
\code{handle_setheader} will first reset all of the currently set headers.
}
\examples{
h <- new_handle()
handle_setopt(h, customrequest = "PUT")
handle_setform(h, a = "1", b = "2")
r <- curl_perform("http://httpbin.org/put", h)
cat(rawToChar(r$content))
}
\seealso{
Other handle.functions: \code{\link{handle_cookies}};
  \code{\link{new_handle}}
}

