% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{ie_proxy}
\alias{ie_proxy}
\alias{ie_proxy_info}
\alias{ie_get_proxy_for_url}
\title{Internet Explorer proxy settings}
\usage{
ie_proxy_info()

ie_get_proxy_for_url(target_url = "http://www.google.com")
}
\arguments{
\item{target_url}{url with host for which to lookup the proxy server}
}
\description{
Lookup and mimic the system proxy settings on Windows as set by Internet
Explorer. This can be used to configure curl to use the same proxy server.
}
\details{
The \code{ie_proxy_info} function looks
up your current proxy settings as configured in IE under "Internet Options"
> "Tab: Connections" > "LAN Settings". The \code{ie_get_proxy_for_url}
determines if and which proxy should be used to connect to a particular
URL. If your settings have an "automatic configuration script" this
involves downloading and executing a PAC file, which can take a while.
}
