% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proxy.R
\name{ie_proxy_info}
\alias{ie_proxy_info}
\title{Internet Explorer Proxy Settings}
\usage{
ie_proxy_info()
}
\description{
This function retrieves the system proxy settings on Windows as
configured in Internet Explorer. This can be used to automatically
configure the libcurl to use the same proxy server.
}

