% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocking.R
\name{mock}
\alias{mock}
\title{Mocking HTTP requests}
\usage{
mock(on = TRUE)
}
\arguments{
\item{on}{(logical) turn mocking on with `TRUE` or turn off with `FALSE`.
By default is `FALSE`}
}
\description{
Mocking HTTP requests
}
