% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{multi}
\alias{multi}
\alias{multi_add}
\alias{multi_remove}
\alias{multi_run}
\title{Async Multi Download}
\usage{
multi_add(handle, complete = NULL, error = NULL)

multi_run(timeout = Inf, total_connetions = 100, host_connections = 6,
  multiplex = TRUE)

multi_remove(handle)
}
\arguments{
\item{handle}{a prepared handle}

\item{complete}{callback on complete}

\item{error}{callback on error}

\item{timeout}{max time in seconds to wait for results. Use \code{0} to poll for results without
waiting at all.}

\item{total_connetions}{limit total concurrent connections}

\item{host_connections}{limit concurrent connections per host}

\item{multiplex}{enable HTTP/2 multiplexing if supported by host and client}
}
\description{
AJAX style concurrent requests, possibly using HTTP/2 multiplexing.
Results are only available via callback functions. Advanced use only!
}
\details{
Requests are created in the usual way using a curl \link{handle} and added
to the scheduler with \link{multi_add}. This function returns immediately
and does not perform the request yet. After zero or more handles have been
added, the \code{multi_run} function performs all requests concurrently.
It returns when all requests have completed or when \code{timeout} has
passed, or when the user interrupts by pressing \code{ESC} or \code{CTRL+C}.
The \code{multi_run} function can be called to resume requests.

A single handle cannot be used for multiple simultaneous requests. However
it is possible to add new requests to a pool while it is running, so you
can re-use a handle within the callback of a request from that same handle.
It is up to the user to make sure the same handle is not used in concurrent
requests.
}
\examples{
h1 <- new_handle(url = "https://httpbin.org/delay/5")
h2 <- new_handle(url = "https://httpbin.org/post", postfields = "bla bla")
h3 <- new_handle(url = "https://urldoesnotexist.xyz")
multi_add(h1, function(res){print(res)})
multi_add(h2, function(res){print(res)})
multi_add(h3)
multi_run(timeout = 3)
}

