% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{multi_async}
\alias{multi_async}
\title{Event-driven Async Requests}
\usage{
multi_async(pool = NULL)
}
\arguments{
\item{pool}{a multi handle created by \link{new_pool}. Default uses a global pool.}
}
\value{
An external pointer to the created thread.
}
\description{
These functions perform all operations to completion asynchronously in a
non-blocking manner using threading.
}
\examples{
\dontshow{if (requireNamespace("later", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
results <- list()
success <- function(x){
  results <<- append(results, list(x))
}
failure <- function(str){
  cat(paste("Failed request:", str, "\n"), file = stderr())
}
# This handle will take longest (2sec)
h1 <- new_handle(url = "https://hb.cran.dev/delay/2")
multi_add(h1, done = success, fail = failure)

# This handle raises an error
h2 <- new_handle(url = "https://urldoesnotexist.xyz")
multi_add(h2, done = success, fail = failure)

# Actually perform the requests
multi_async()

# Check the results (after waiting for completion)
Sys.sleep(3)
later::run_now()
results
\dontshow{\}) # examplesIf}
}
\seealso{
Multi interface: \link{multi}
}
