% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{multipart}
\alias{multipart}
\alias{form_file}
\alias{form_data}
\title{POST files or data}
\usage{
form_file(path, type = NULL)

form_data(value, type = NULL)
}
\arguments{
\item{path}{a string with a path to an existing file on disk}

\item{type}{MIME content-type of the file.}

\item{value}{a character or raw vector to post}
}
\description{
Build multipart form data elements. The \code{form_file} function uploads a
file. The \code{form_data} function allows for posting a string or raw vector
with a custom content-type.
}
