% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/handle.R
\name{new_handle}
\alias{new_handle}
\title{Create new libcurl handle}
\usage{
new_handle(...)
}
\arguments{
\item{...}{Initial options to set on handle. See \code{\link{handle_setopt}}
for more details.}
}
\value{
A handle object, an external pointer to the underlying
  Curl handle.
}
\description{
Create new libcurl handle
}
\seealso{
Other handle.functions: \code{\link{handle_cookies}};
  \code{\link{handle_reset}}, \code{\link{handle_setform}},
  \code{\link{handle_setheaders}},
  \code{\link{handle_setopt}}
}

