% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nslooup.R
\name{nslookup}
\alias{nslookup}
\title{Lookup a hostname}
\usage{
nslookup(host)
}
\arguments{
\item{hostname}{a string with a host}
}
\description{
The same thing as \link{nsl} but supports IPv6 and works on all platforms.
}

