% Generated by roxygen2 (4.0.2): do not edit by hand
\name{utility functions}
\alias{curl_escape}
\alias{curl_options}
\alias{curl_unescape}
\alias{curl_version}
\alias{utility functions}
\title{Curl utility functions}
\usage{
curl_escape(url)

curl_unescape(url)

curl_version()

curl_options()
}
\arguments{
\item{url}{a string (typically a url or parameter) to be URL encoded}
}
\description{
Utility functions for working with curl.
}
\examples{
curl_version()
out <- curl_escape("foo = bar + 5")
curl_unescape(out)
}

