% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curl.R
\name{curl}
\alias{curl}
\title{Curl connection interface}
\usage{
curl(url = "https://hb.cran.dev/get", open = "", handle = new_handle())
}
\arguments{
\item{url}{character string. See examples.}

\item{open}{character string. How to open the connection if it should be opened
initially. Currently only "r" and "rb" are supported.}

\item{handle}{a curl handle object}
}
\description{
Drop-in replacement for base \code{\link{url}} that supports https, ftps,
gzip, deflate, etc. Default behavior is identical to \code{\link{url}}, but
request can be fully configured by passing a custom \code{\link{handle}}.
}
\details{
As of version 2.3 curl connections support \code{open(con, blocking = FALSE)}.
In this case \code{readBin} and \code{readLines} will return immediately with data
that is available without waiting. For such non-blocking connections the caller
needs to call \code{\link{isIncomplete}} to check if the download has completed
yet.
}
