% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{curl_echo}
\alias{curl_echo}
\alias{find_port}
\title{Echo Service}
\usage{
curl_echo(handle, port = find_port(), progress = interactive(), file = NULL)

find_port(range = NULL)
}
\arguments{
\item{handle}{a curl handle object}

\item{port}{the port number on which to run httpuv server}

\item{progress}{show progress meter during http transfer}

\item{file}{path or connection to write body. Default returns body as raw vector.}

\item{range}{optional integer vector of ports to consider}
}
\description{
This function is only for testing purposes. It starts a local httpuv server to
echo the request body and content type in the response.
}
