% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/perform.R
\name{curl_perform}
\alias{curl_perform}
\title{Perform a request}
\usage{
curl_perform(url, handle = new_handle())
}
\arguments{
\item{url}{A character string naming the URL of a resource to be downloaded.}

\item{handle}{a curl handle object}
}
\description{
Perform a request
}
\examples{
# Redirect + cookies
res <- curl_perform("http://httpbin.org/cookies/set?foo=123&bar=ftw")
}

