% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp.R
\name{curl_upload}
\alias{curl_upload}
\title{Upload a File}
\usage{
curl_upload(file, url, verbose = TRUE, reuse = TRUE, ...)
}
\arguments{
\item{file}{connection object or path to an existing file on disk}

\item{url}{where to upload, should start with e.g. \code{ftp://}}

\item{verbose}{emit some progress output}

\item{reuse}{try to keep alive and recycle connections when possible}

\item{...}{other arguments passed to \code{\link{handle_setopt}}, for
example a \code{username} and \code{password}.}
}
\description{
Upload a file to e.g. an \code{ftp://} or \code{sftp://} (ssh) server. Note
that sftp is only available for libcurl clients built with libssh2.
}
\examples{
\donttest{# Upload package to winbuilder:
curl_upload('mypkg_1.3.tar.gz', 'ftp://win-builder.r-project.org/R-devel/')
}
}
