% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/handle.R
\name{new_handle}
\alias{get_handle_cookies}
\alias{new_handle}
\title{Create new libcurl handle}
\usage{
new_handle()

get_handle_cookies(handle)
}
\arguments{
\item{handle}{a curl handle object}
}
\description{
This function creates a new libcurl easy handle object.
}
\examples{
h <- new_handle()
req <- curl_perform("http://httpbin.org/cookies/set?foo=123&bar=ftw", handle = h)

# See http://www.cookiecentral.com/faq/#3.5
get_handle_cookies(h)
}

