% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/handle.R
\name{new_handle}
\alias{handle_cookies}
\alias{handle_reset}
\alias{handle_setform}
\alias{handle_setheaders}
\alias{handle_setopt}
\alias{new_handle}
\title{Create new libcurl handle}
\usage{
new_handle(...)

handle_setopt(handle, ...)

handle_reset(handle)

handle_setheaders(handle, ...)

handle_setform(handle, ...)

handle_cookies(handle)
}
\arguments{
\item{...}{arguments passed to handle_setopt}

\item{handle}{a curl handle object}

\item{...}{additional options / headers to be set in the handle.}
}
\description{
Functions to create and manipualte handle objects. Note that currently
handle_setopt will append options whereas handle_setheader will reset
all of the currently set headers.
}
\examples{
h <- new_handle()
handle_cookies(h)

# Server sets cookies
req <- curl_perform("http://httpbin.org/cookies/set?foo=123&bar=ftw", handle = h)
handle_cookies(h)

# Server deletes cookies
req <- curl_perform("http://httpbin.org/cookies/delete?foo", handle = h)
handle_cookies(h)
}

