% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/handle.R
\name{handle}
\alias{handle}
\alias{new_handle}
\title{Create new libcurl handle}
\usage{
new_handle()
}
\description{
Functions to create and manipualte handle objects. Note that currently
handle_setopt will append options whereas handle_setheader will reset
all of the currently set headers.
}
\examples{
h <- new_handle()
h$cookies()

# Server sets cookies
req <- curl_perform("http://httpbin.org/cookies/set?foo=123&bar=ftw", handle = h)
h$cookies()

# Server deletes cookies
req <- curl_perform("http://httpbin.org/cookies/delete?foo", handle = h)
h$cookies()
}

