% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proxy.R
\name{ie_proxy_info}
\alias{ie_proxy_info}
\title{Internet Explorer proxy settings}
\usage{
ie_proxy_info()
}
\description{
This function retrieves the system proxy settings on Windows as
set in Internet Explorer. This can be used to configure curl to
use the same proxy server.
}

