% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{multi_async}
\alias{multi_async}
\title{Event-driven Async Requests}
\usage{
multi_async(pool)
}
\arguments{
\item{pool}{A multi handle created by \link{new_pool}. This must be supplied,
and there is no default.}
}
\value{
An external pointer to the created thread.
}
\description{
\code{multi_async} is a threaded function that performs all operations to
completion asynchronously in a non-blocking manner.
}
\details{
This function does not default to the global pool. This helps to ensure that
other functions are not unintentionally called on it concurrently, as it is
not safe to do so.

Once this function is called, further actions must not be made on the
supplied pool until all async operations have completed.
}
\examples{
\dontshow{if (requireNamespace("later", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
success <- function(x){
results <<- append(results, list(x))
}

failure <- function(str){
cat(paste("Failed request:", str, "\n"), file = stderr())
}

results <- list()
pool <- curl::new_pool()

# Create handles
h1 <- new_handle(url = "https://hb.cran.dev/delay/2")
h2 <- new_handle(url = "https://hb.cran.dev/delay/1")

# Add handles:
multi_add(h1, done = success, fail = failure, pool = pool)
multi_add(h2, done = success, fail = failure, pool = pool)

# Actually perform the requests
multi_async(pool = pool)

# Wait for completion
Sys.sleep(3)
later::run_now()

# check all results are present
length(results)
\dontshow{\}) # examplesIf}
}
\seealso{
Multi interface: \link{multi}
}
