% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_download.R
\name{multi_download}
\alias{multi_download}
\title{Download multiple files concurrently}
\usage{
multi_download(
  urls,
  destfiles = NULL,
  resume = FALSE,
  progress = FALSE,
  timeout = Inf,
  ...
)
}
\arguments{
\item{urls}{vector with files to download}

\item{destfiles}{vector (of equal length as \code{urls}) with paths of output files,
or \code{NULL} to use \link{basename} of urls.}

\item{resume}{if the file already exists, resume the download. Note that this may
change server responses, see details.}

\item{progress}{print download progress information}

\item{timeout}{in seconds, passed to \link{multi_run}}

\item{...}{extra handle options passed to each request \link{new_handle}}
}
\description{
Wrapper for \code{\link[=multi_run]{multi_run()}} to download multiple requests concurrently. Also supports
resuming downloads for large files. This function does not error in case any of the
requests fail; you should inspect the returned results to see which of the requests
were completed successfully.
}
\details{
This function returns a data frame with information about each request. The \code{success}
column indicates if the request was successfully completed (regardless of the http
status code). If it failed, e.g. due to a connection issue, the error message is in
the \code{error} column. A \code{success} value \code{NA} means that the download was still in
progress when the function reached the elapsed \code{timeout} and has not yet been
completed, perhaps it could be resumed.

Upon completion, it is important to check the \code{status_code} column to see if any of
the request may have had an HTTP error, and hence the downloaded file contains an
error page instead of the requested content. Note that when \code{resume = TRUE} you can
expect HTTP-206 or HTTP-416 responses. The latter could indicate that the file was
already downloaded completely, hence there was no content left to resume.
}
\examples{
urls <- c('https://cran.r-project.org/src/contrib/Archive/V8/V8_4.2.1.tar.gz',
'https://cran.r-project.org/src/contrib/Archive/curl/curl_4.3.2.tar.gz',
'https://urldoesnotexist.xyz/nothing.zip',
'https://github.com/jeroen/curl/archive/refs/heads/master.zip',
'https://httpbin.org/status/418')

multi_download(urls)
}
