% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_download.R
\name{multi_download}
\alias{multi_download}
\title{Download multiple files concurrently}
\usage{
multi_download(urls, destfiles = NULL, resume = FALSE, timeout = Inf, ...)
}
\arguments{
\item{urls}{vector with files to download}

\item{destfiles}{vector (of equal length as `urls`) with paths of output files,
or `NULL` to use [basename] of urls.}

\item{resume}{if the file already exists, resume the download.}

\item{timeout}{in seconds, passed to [multi_run]}

\item{...}{extra handle options passed to [new_handle]}
}
\description{
Wrapper for [multi_run()] to download multiple requests concurrently. Also supports
resuming downloads for large files. This function does not error in case any of the
requests fail; instead it returns a data frame with information about the requests.
}
\examples{
urls <- c('https://cran.r-project.org/src/contrib/Archive/V8/V8_4.2.1.tar.gz',
'https://cran.r-project.org/src/contrib/Archive/curl/curl_4.3.2.tar.gz',
'https://urldoesnotexist.xyz/nothing.zip',
'https://httpbin.org/status/418')

multi_download(urls)
}
