% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nslooup.R
\name{nslookup}
\alias{nslookup}
\title{Lookup a hostname}
\usage{
nslookup(host, error = TRUE)
}
\arguments{
\item{host}{a string with a hostname}

\item{error}{raise an error for failed DNS lookup. Otherwise returns \code{NULL}.}
}
\description{
Similar to \link{nsl} but supports IPv6 and works on all platforms.
Default behavior raises an error if lookup fails.
}
\examples{
nslookup("www.r-project.org")
nslookup("ipv6.test-ipv6.com")
}

