% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nslookup.R
\name{has_internet}
\alias{has_internet}
\title{Lookup a hostname}
\usage{
has_internet()
}
\arguments{
\item{host}{a string with a hostname}

\item{error}{raise an error for failed DNS lookup. Otherwise returns \code{NULL}.}

\item{ipv4_only}{always return ipv4 address. Set to \code{FALSE} to allow for ipv6 as well.}

\item{multiple}{returns multiple ip addresses if possible}
}
\description{
The \code{nslookup} function is similar to \code{nsl} but works on all platforms
and can resolve ipv6 addresses if supported by the OS. Default behavior raises an
error if lookup fails.
}
\details{
The \code{has_internet} function tests for internet connectivity by performing a
dns lookup. If a proxy server is detected, it will also check for connectivity by
connecting via the proxy.
}
