% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_url}
\alias{parse_url}
\title{Parse a URL}
\usage{
parse_url(url)
}
\arguments{
\item{url}{a character string of length one}
}
\description{
Wrapper for the libcurl \href{https://curl.se/libcurl/c/libcurl-url.html}{URL parsing interface}.
This requires at least libcurl 7.62.
}
