% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{send_mail}
\alias{send_mail}
\title{Send email}
\usage{
send_mail(mail_from, mail_rcpt, message, smtp_server = "localhost",
  use_ssl = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{mail_from}{email address of the sender.}

\item{mail_rcpt}{one or more recipient email addresses. Do not include names,
these go into the \code{message} headers.}

\item{message}{either a string or connection with (properly formatted) email
message, including sender/recipient/subject headers. See example.}

\item{smtp_server}{hostname or address of the SMTP server, or, an
\code{smtp://} or \code{smtps://} URL. See "Specifying the server, port,
and protocol" below.}

\item{use_ssl}{Attempt to encrypt the connection via STARTTLS, if the server
supports it. Defaults to \code{TRUE} and should only be set to \code{FALSE}
if you have a specific reason to disable STARTTLS.}

\item{verbose}{print output}

\item{...}{other options passed to \code{\link{handle_setopt}}. In most cases
you will need to set a \code{username} and \code{password} to authenticate
with the SMTP server.}
}
\description{
Use the curl SMTP client to send an email. The \code{message} argument must be
properly formatted \href{https://tools.ietf.org/html/rfc2822}{RFC2822} email message with From/To/Subject headers and CRLF
line breaks.
}
\details{

}
\section{Specifying the server, port, and protocol}{


The \code{smtp_server} argument takes a hostname, or an SMTP URL:

\itemize{
  \item \code{mail.example.com} - hostname only
  \item \code{mail.example.com:587} - hostname and port
  \item \code{smtp://mail.example.com} - protocol and hostname
  \item \code{smtp://mail.example.com:587} - full SMTP URL
  \item \code{smtps://mail.example.com:465} - full SMTPS URL
}

By default, the port will be 25, unless \code{smtps://} is specified--then
the default will be 465 instead.
}

\section{Encrypting connections via SMTPS or STARTTLS}{


SMTP connections can be encrypted in one of two ways.

If your email server listens on port 465, then use an
\code{smtps://hostname:465} URL. The SMTPS protocol \emph{guarantees} that
TLS will be used to protect your communications.

If your email server listens on port 25 or 587, use an \code{smtp://} URL and
\code{use_ssl = TRUE} (the default). This uses STARTTLS to
\emph{opportunistically} use encryption (i.e. encryption will only be used if
the server supports it).
}

\examples{
\donttest{# Message in RFC2822 format
message <- 'From: "Testbot" <jeroen@ocpu.io>
To: "Jeroen Ooms" <jeroenooms@gmail.com>
Subject: Hello there!

Hi Jeroen,
I am sending this email using curl.'


# Actually send the email
recipients <- c('test@opencpu.org', 'jeroenooms@gmail.com')
sender <- 'test@ocpu.io'
username <- 'mail@ocpu.io'
password <- askpass::askpass(paste("SMTP server password for", username))
send_mail(sender, recipients, smtp_server = 'smtp.mailgun.org',
  message = message, username = username, password = password)}
}
